+++ lib/Models/Terria.ts
@@ -57,6 +57,7 @@ import Workbench from "./Workbench";
 import openGroup from "./openGroup";
 import getDereferencedIfExists from "../Core/getDereferencedIfExists";
 import SplitItemReference from "./SplitItemReference";
+// import overrides from "../Overrides/defaults.jsx";
 
 interface ConfigParameters {
   [key: string]: ConfigParameters[keyof ConfigParameters];
@@ -87,6 +88,7 @@ interface ConfigParameters {
   locationSearchBoundingBox?: number[];
   googleAnalyticsKey?: string;
   rollbarAccessToken?: string;
+  showWelcomeMessage?: boolean;
   helpContent?: any[];
 }
 
@@ -128,6 +130,7 @@ export default class Terria {
   readonly overlays = new Workbench();
   readonly catalog = new Catalog(this);
   readonly timelineClock = new Clock({ shouldAnimate: false });
+  // readonly overrides: any = overrides; // TODO: add options.functionOverrides like in master
 
   @observable
   readonly mainViewer = new TerriaViewer(
@@ -195,6 +198,7 @@ export default class Terria {
     locationSearchBoundingBox: undefined,
     googleAnalyticsKey: undefined,
     rollbarAccessToken: undefined,
+    showWelcomeMessage: false,
     helpContent: []
   };