+++ lib/ReactViews/StandardUserInterface/StandardUserInterface.jsx
@@ -28,7 +28,7 @@ import StoryBuilder from "./../Story/StoryBuilder.jsx";
 
 import SatelliteGuide from "../Guide/SatelliteGuide";
 import TourPortal from "../Tour/TourPortal";
-// import WelcomeMessage from "../WelcomeMessage/WelcomeMessage";
+import WelcomeMessage from "../WelcomeMessage/WelcomeMessage";
 
 import { Small, Medium } from "../Generic/Responsive";
 import classNames from "classnames";
@@ -194,7 +194,7 @@ const StandardUserInterface = observer(
           />
           <TourPortal terria={terria} viewState={this.props.viewState} />
           <div className={Styles.storyWrapper}>
-            {/* <WelcomeMessage viewState={this.props.viewState} /> */}
+            <WelcomeMessage viewState={this.props.viewState} />
             <div
               className={classNames(Styles.uiRoot, {
                 [Styles.withStoryBuilder]: showStoryBuilder