+++ lib/Styled/Text.jsx
@@ -108,6 +108,11 @@ export const Text = styled.div`
     font-size: 26px;
     line-height: 32px;
   `}
+
+  ${props => props.styledSize && `font-size: ${props.styledSize}`};
+  ${props =>
+    props.styledLineHeight && `line-height: ${props.styledLineHeight}`};
+  
   ${props =>
     props.highlightLinks &&
     `